+++ lib/src/core/domain/user.dart
@@ -1,3 +1,5 @@
+import 'package:where_to_go_today/src/core/domain/place.dart';
+
 /// Модель пользователя
 class User {
   final int id;
@@ -7,6 +9,7 @@ class User {
   final String email;
   final String birthDate;
   final String photoUrl;
+  final List<Place> favoritePlaces;
 
   User({
     required this.id,
@@ -16,5 +19,6 @@ class User {
     required this.email,
     required this.birthDate,
     this.photoUrl = '',
+    this.favoritePlaces = const [],
   });
 }