+++ lib/src/features/main/profile/profile_vm.g.dart
@@ -9,6 +9,63 @@ part of 'profile_vm.dart';
 // ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic
 
 mixin _$ProfileVm on _ProfileVm, Store {
+  final _$_ProfileVmActionController = ActionController(name: '_ProfileVm');
+
+  @override
+  void onAvatar() {
+    final _$actionInfo =
+        _$_ProfileVmActionController.startAction(name: '_ProfileVm.onAvatar');
+    try {
+      return super.onAvatar();
+    } finally {
+      _$_ProfileVmActionController.endAction(_$actionInfo);
+    }
+  }
+
+  @override
+  void onEditUserDetails() {
+    final _$actionInfo = _$_ProfileVmActionController.startAction(
+        name: '_ProfileVm.onEditUserDetails');
+    try {
+      return super.onEditUserDetails();
+    } finally {
+      _$_ProfileVmActionController.endAction(_$actionInfo);
+    }
+  }
+
+  @override
+  void onSettings() {
+    final _$actionInfo =
+        _$_ProfileVmActionController.startAction(name: '_ProfileVm.onSettings');
+    try {
+      return super.onSettings();
+    } finally {
+      _$_ProfileVmActionController.endAction(_$actionInfo);
+    }
+  }
+
+  @override
+  void onFavoritePlace(int id) {
+    final _$actionInfo = _$_ProfileVmActionController.startAction(
+        name: '_ProfileVm.onFavoritePlace');
+    try {
+      return super.onFavoritePlace(id);
+    } finally {
+      _$_ProfileVmActionController.endAction(_$actionInfo);
+    }
+  }
+
+  @override
+  void onProfileSignOut() {
+    final _$actionInfo = _$_ProfileVmActionController.startAction(
+        name: '_ProfileVm.onProfileSignOut');
+    try {
+      return super.onProfileSignOut();
+    } finally {
+      _$_ProfileVmActionController.endAction(_$actionInfo);
+    }
+  }
+
   @override
   String toString() {
     return '''