+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - feat: SignInScreenVm - add phone authorization
 - feat: OnboardingScreen
 - feat: MainScreen + PlacesScreen
+- feat: profile screen
 
 ## 0.3.0